week8/pyproject.toml
[tool.poetry]
name = "week8"
version = "0.1.0"
description = ""
authors = ["AliceKuo <ming-chieh.kuo@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
mysql-connector-python = "^8.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
